+++ sqllogictest/src/runner.rs
@@ -642,14 +642,12 @@ impl<D: AsyncDB, M: MakeConnection<Conn = D>> Runner<D, M> {
                     cmd.arg("-c").arg(command);
                     cmd
                 };
-                if expected_stdout.is_some() {
-                    // only capture stdout if there's an expected value
-                    cmd.stdout(std::process::Stdio::piped());
-                }
+                cmd.stdout(std::process::Stdio::piped());
+                cmd.stderr(std::process::Stdio::piped());
                 let result = D::run_command(cmd).await;
                 #[derive(thiserror::Error, Debug)]
-                #[error("process exited unsuccessfully: {0}")] // message from unstable `ExitStatusError`
-                struct SystemError(ExitStatus);
+                #[error("process exited unsuccessfully: {0}\nstdout: {1}\nstderr: {2}")]
+                struct SystemError(ExitStatus, String, String);
 
                 let mut stdout = None;
                 let error: Option<AnyError> = match result {
@@ -660,7 +658,11 @@ impl<D: AsyncDB, M: MakeConnection<Conn = D>> Runner<D, M> {
                             }
                             None
                         } else {
-                            Some(Arc::new(SystemError(output.status)))
+                            Some(Arc::new(SystemError(
+                                output.status,
+                                String::from_utf8_lossy(&output.stdout).to_string(),
+                                String::from_utf8_lossy(&output.stderr).to_string(),
+                            )))
                         }
                     }
                     Err(e) => Some(Arc::new(e)),