+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [0.20.0] - 2024-04-08
 
+* Show stdout, stderr when `system` command fails.
 * Support matching stdout for `system`
   ```
   system ok